I abstract more into libs.
