+++ hooks/useTBAPurchase.tsx
@@ -1,22 +1,9 @@
-import { BigNumber } from "ethers"
 import { useState } from "react"
 import multicallAbi from "@/lib/abi/multicall3.json"
-import getMintMulticallCalls from "@/lib/getMintMulticallCalls"
-import {
-  CATALOG_PRICE,
-  CHAIN_ID,
-  MULTICALL3_ADDRESS,
-  SPOTIFY_DROP_ADDRESS,
-  ZORA_PRICE,
-} from "@/lib/consts"
+import { CHAIN_ID, MULTICALL3_ADDRESS } from "@/lib/consts"
 import handleTxError from "@/lib/handleTxError"
 import { useUserProvider } from "@/providers/UserProvider"
-import getZoraNextTokenId from "@/lib/getZoraNextTokenId"
-import getAccount from "@/lib/tokenbound/getAccount"
-import getCosignMintCall from "@/lib/getCosignMintCall"
-import getTBAInitializeCall from "@/lib/getTBAInitializeCall"
-import getSoundMintCall from "@/lib/sound/getSoundMintCall"
-import getSoundInterstellarMintCall from "@/lib/sound/getSoundInterstellarMintCall"
+import getPreparedMulticalls from "@/lib/getPreparedMulticalls"
 import useConnectedWallet from "./useConnectedWallet"
 import usePrivySendTransaction from "./usePrivySendTransaction"
 import useWalletTransaction from "./useWalletTransaction"
@@ -36,47 +23,12 @@ const useTBAPurchase = () => {
       if (!connectedWallet) return false
 
       setLoading(true)
-      const zoraTotalPrice = BigNumber.from(ZORA_PRICE).mul(1)
-      const zoraNextTokenId = await getZoraNextTokenId()
-      const zoraQuantity = 1
-      const tbaCalls = getMintMulticallCalls(
-        zoraNextTokenId,
-        connectedWallet as string,
-        zoraQuantity,
-        zoraTotalPrice.toString(),
-      ) as any
-      const tba = getAccount(zoraNextTokenId)
-      const tbaInitializationCall = getTBAInitializeCall(tba)
-      const soundMintCall = await getSoundMintCall(tba, CHAIN_ID, SPOTIFY_DROP_ADDRESS)
-
-      if (!soundMintCall) {
+      const prepared = await getPreparedMulticalls(connectedWallet as `0x${string}`)
+      if (!prepared) {
         setLoading(false)
         return false
       }
-      const interstellarMintCall = await getSoundInterstellarMintCall(
-        tba,
-        connectedWallet as `0x${string}`,
-      )
-
-      console.log("SWEETS soundMintCall", soundMintCall)
-      console.log("SWEETS bridgeCall same as soundMintCall?", interstellarMintCall)
-
-      const soundMintCallValue = BigNumber.from(soundMintCall.value)
-      const cosignMintCall = getCosignMintCall(tba)
-      const cosignMintValue = BigNumber.from(CATALOG_PRICE)
-      const soundBridgeValue = BigNumber.from(interstellarMintCall.value)
-      const totalPrice = soundMintCallValue
-        .add(cosignMintValue)
-        .add(zoraTotalPrice)
-        .add(soundBridgeValue)
-      const hexValue = totalPrice.toHexString()
-      const calls = [
-        ...tbaCalls,
-        tbaInitializationCall,
-        soundMintCall,
-        cosignMintCall,
-        interstellarMintCall,
-      ]
+      const { calls, hexValue } = prepared
 
       if (isLoggedByEmail) {
         const response = await sendTxByPrivy(
